move PlayersTable to QueueSection folder
