+++ src/pages/Lobby/components/QueueSection/PlayersTable/styles.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { theme } from '../../../../shared/theme';
+import { theme } from '@shared/theme';
 import { RowProps } from './types';
 import { css } from '@emotion/react';
